requirements/local.txt
asgiref==3.7.2
Django==4.2.2
gunicorn==20.1.0
tzdata==2022.1
sqlparse==0.4.4
loguru==0.7.0
python-environ==0.4.54
django-phonenumber-field==7.1.0
django-taggit==3.1.0
pandas
scikit-learn
nltk
pickle5
numpy